requirements.txt
google-generativeai
PyPDF2
mysql-connector-python
openai
python-dotenv
